seperated logic part
