+++ src/componants/order/CreateOrder/CreateOrder.jsx
@@ -10,6 +10,10 @@ import SuccessOrder from "./SuccessOrder/SuccessOrder";
 import { useReducer } from "react";
 import FooterOrder from "../footer/footer";
 
+//------------operations------------//
+import { completeDataOperation } from "./operation/completeDataOperation";
+import { productReducer } from "./operation/productReducer";
+
 export default function CreateOrder(){
 
     const navigate = useNavigate();
@@ -22,176 +26,40 @@ export default function CreateOrder(){
 
     const[completeData,setCompleteData] = useState(initialData);
 
-    const[error,setError] = useState(false);
-
     function completeDataUpdate(productData){
-        setCompleteData(prev=>{
-            let temp = {...prev};
-            let proData = [...productData];
-            
-
-
-            temp.orders=[];
-            temp.total_price = 90;
-            proData.map(data=>{
-
-                if(data.quantity > 0 && data.price > 0){
-
-                    let washType = "";
-
-                     
-                        if( data.washing_types.pack.status){
-                            washType = data.washing_types.pack.type + "," + washType;
-                            
-                        }
-                        if(data.washing_types.fold.status){
-                            washType = data.washing_types.fold.type + "," + washType;
-                            
-                        }
-                        if(data.washing_types.press.status){
-                            washType = data.washing_types.press.type + "," + washType;
-                            
-                        }
-                        if(data.washing_types.wash.status){
-                            washType = data.washing_types.wash.type + "," + washType;
-                            
-                        }
-  
-
-                     temp.orders.push({
-                        product_type: data.product_type,
-                        quantity:data.quantity,
-                        wash_type:washType,
-                        price:data.price
-                     })
-
-                    temp.total_price = temp.total_price + (data.price*data.quantity);
-                }
-            })
-
-             
-
-            return temp;
-        })
+        completeDataOperation(productData,setCompleteData);
     }
 
+    
+    const[error,setError] = useState(false);
+    
+    
     // console.log(completeData);
-
+    
     const[success, setSuccess] = useState(false);
-
+    
     function handleSucess(condition){
         setSuccess(condition);
     }
-
+    
     const [summaryTg,setSummaryTg] = useState(false);
-
+    
     function summaryHandler(condition){
         setSummaryTg(condition);
     }
+    
 
-    function productReducer(state,action){
-        switch(action.type){
-            case "initLoad":{
-                return action.payload.data;
-            }
-            case "quantity":{
-                let temp = [...state];
-                temp.map(data=>{
-                    if(data.id === action.payload.id){
-                        data.quantity = action.payload.value;
-                    }
-                })
-                return temp;
-            }
-            case "wash":{
-                let temp =[...state]
-                temp.map(data=>{
-                    if(data.id === action.payload.id){
-                        data.washing_types.wash.status = !data.washing_types.wash.status;
-                        if(data.washing_types.wash.status){
-                            data.price = data.price + action.payload.price;  
-                        }else{
-                            data.price = data.price - action.payload.price;
-                        }
-                    }
-                })
-                return temp;
-            }
-            case "fold":{
-                let temp =[...state]
-                temp.map(data=>{
-                    if(data.id === action.payload.id){
-                        data.washing_types.fold.status = !data.washing_types.fold.status;
-                        if(data.washing_types.fold.status){
-                            data.price = data.price + action.payload.price;
-                        }else{
-                            data.price = data.price - action.payload.price;
-
-                        }
-                    }
-                })
-                return temp;
-            }
-            case "press":{
-                let temp =[...state]
-                temp.map(data=>{
-                    if(data.id === action.payload.id){
-                        data.washing_types.press.status = !data.washing_types.press.status;
-                        if(data.washing_types.press.status){
-
-                            data.price = data.price + action.payload.price;
-                        }else{
-                            data.price = data.price - action.payload.price;
-                        }
-                    }
-                })
-                return temp;
-            }
-            case "pack":{
-                let temp =[...state]
-                temp.map(data=>{
-                    if(data.id === action.payload.id){
-                        data.washing_types.pack.status = !data.washing_types.pack.status;
-                        if(data.washing_types.pack.status){
-
-                            data.price = data.price + action.payload.price;
-                        }else{
-                            data.price = data.price - action.payload.price;
-
-                        }
-                    }
-                })
-                return temp;
-            }
-            case "reset":{
-                let temp = [...state];
-                temp.map(data=>{
-                    if(data.id === action.payload.id){
-                        data.quantity = 0;
-                        data.price = 0;
-                        data.washing_types.wash = {...data.washing_types.wash,status : false};
-                        data.washing_types.press = {...data.washing_types.press,status : false};
-                        data.washing_types.fold = {...data.washing_types.fold,status : false};
-                        data.washing_types.pack = {...data.washing_types.pack,status : false};
-                    }
-                })
-
-                return temp;
-            }
-            default:{
-                return state;
-            }
-        }
-
-    }
     
     
     const [productData,productDispatcher] = useReducer(productReducer,[]);
+    
+    
+
     // console.log(productData);
     useEffect(()=>{
         const url = process.env.REACT_APP_API +"/product";
-         async function FCall(){
-           try{
+        async function FCall(){
+            try{
               const response = await axios(url,{
                   headers:{
                       authorization: localStorage.getItem('token'),
